messed with music a bit
