+++ HAGJ-Demo/Assets/Audio/900 Music/Music - Battle Choral Full 118bpm -loop.wav.meta
@@ -4,7 +4,7 @@ AudioImporter:
   externalObjects: {}
   serializedVersion: 6
   defaultSettings:
-    loadType: 0
+    loadType: 1
     sampleRateSetting: 0
     sampleRateOverride: 44100
     compressionFormat: 1
