--- HAGJ-Demo/Assets/MusicPlayback.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb27f691abffeb244837b93c4d8f0819
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
