+++ HAGJ-Demo/UserSettings/EditorUserSettings.asset
@@ -27,6 +27,9 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c7e05121e176a002924707f7a5065acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c72111f195807292633707c67083debf42d
+      flags: 0
+    RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c7e05121e176a002924707f7a5065a2b27a2decee22f0
       flags: 0
     vcSharedLogLevel:
